requirements.txt
PyMuPDF==1.24.9
attrs==24.2.0
azure-core==1.30.1
azure-storage-blob==12.19.1
backcall==0.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.2.0
blinker==1.7.0
click==8.1.7
decorator==4.4.2
defusedxml==0.7.1
dnspython==1.16.0
docopt==0.6.2
eventlet==0.31.1
fastjsonschema==2.20.0
ffmpeg-python==0.2.0
Flask==3.0.3
flask_sqlalchemy==3.0.5
flask-bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-SocketIO==5.3.6
future==1.0.0
Flask==3.0.3
PyJWT==2.9.0
google-auth==2.35.0
greenlet==3.0.3
gunicorn==22.0.0
isodate==0.6.1
itsdangerous==2.1.2
beautifulsoup4==4.12.3
requests==2.31.0
Jinja2==3.1.3
jsonpointer==2.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyterlab_pygments==0.3.0
M2Crypto==0.41.0
MarkupSafe==2.1.5
mistune==3.0.2
mkl-service==2.4.0
moviepy==1.0.3
munkres==1.1.4
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
pandocfilters==1.5.1
Passbook==1.0.2
pickleshare==0.7.5
pipreqs==0.5.0
pgvector==0.3.5
proglog==0.1.10
psycopg2-binary==2.9.9
pypng==0.20220715.0
python-dotenv==1.0.1
python-engineio==4.9.1
python-http-client==3.3.7
python-socketio==5.11.3
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==11.0.0
openai==1.54.3
pydantic==1.10.9
qrcode==7.4.2
referencing==0.35.1
rpds-py==0.21.0
sendgrid==6.11.0
simple-websocket==1.0.0
soupsieve==2.6
starkbank-ecdsa==2.2.0
tinycss2==1.4.0
typing_extensions==4.11.0
wallet==2.1.525
webencodings==0.5.1
Werkzeug==3.0.2
wsproto==1.2.0
yarg==0.1.9
SQLAlchemy==2.0.36
PyJWT==2.9.0
azure-storage-blob==12.19.1
Werkzeug==3.0.2
ipinfo
pinecone-client
